code refactor for base setup of v2
