+++ src/App.js
@@ -1,61 +1,23 @@
 import React from "react";
-import { useGlobalContext } from "./context";
+import { useGlobalContext } from "./context/context";
 
-import SetupForm from "./SetupForm";
-import Loading from "./Loading";
-import Modal from "./Modal";
+import SetupForm from "./components/SetupForm";
+import Loading from "./components/Loading";
+import Modal from "./components/Modal";
+import Quiz from "./components/Quiz";
 function App() {
-  const {
-    waiting,
-    isLoading,
-    questions,
-    index,
-    correct,
-    nextQuestion,
-    checkAnswer,
-  } = useGlobalContext();
+  const { waiting, isLoading } = useGlobalContext();
   if (waiting) {
     return <SetupForm />;
   }
   if (isLoading) {
     return <Loading />;
   }
-  const { question, incorrect_answers, correct_answer } = questions[index];
 
-  let answers = [...incorrect_answers];
-  let tempIndex = Math.floor(Math.random() * 4);
-  if (tempIndex === 3) {
-    answers.push(correct_answer);
-  } else {
-    answers.push(answers[tempIndex]);
-    answers[tempIndex] = correct_answer;
-  }
   return (
     <main>
       <Modal />
-      <section className="quiz">
-        <p className="correct-answers">
-          correct answers : {correct}/ {index + 1}
-        </p>
-        <article className="container">
-          <h2 dangerouslySetInnerHTML={{ __html: question }} />
-          <div className="btn-container">
-            {answers.map((answer, index) => {
-              return (
-                <button
-                  key={index}
-                  className="answer-btn"
-                  dangerouslySetInnerHTML={{ __html: answer }}
-                  onClick={() => checkAnswer(correct_answer === answer)}
-                />
-              );
-            })}
-          </div>
-        </article>
-        <button className="next-question" onClick={nextQuestion}>
-          next question
-        </button>
-      </section>
+      <Quiz />
     </main>
   );
 }
