+++ src/components/Modal.js
@@ -1,5 +1,5 @@
 import React from "react";
-import { useGlobalContext } from "./context";
+import { useGlobalContext } from "../context/context";
 
 const Modal = () => {
   const { isModalOpen, closeModal, correct, questions } = useGlobalContext();
